Moved Game object from ThrillerTest.py (temporary place) to GameObject.py
