+++ GameObject.py
@@ -1,3 +1,60 @@
+class Game:
+   def __init__( self, world, inventory, actions ):
+      self.world = world
+      self.inventory = inventory
+      self.actions = actions
+
+   def move_between_entities( self, name, from_entity, to_entity ):
+      object = from_entity.take( name )
+      if ( not object is None ):
+         to_entity.put( object )
+         return object
+      return None
+
+   def take( self, name ):
+      return self.move_between_entities( name, self.world, self.inventory )
+
+   def drop( self, name ):
+      return self.move_between_entities( name, self.inventory, self.world )
+
+   def destroy( self, name ):
+      self.world.take( name )
+      self.inventory.take( name )
+      return None 
+
+   def has( self, name ):
+      object = self.inventory.find( name )
+      return object
+
+   def find( self, name ):
+      object = self.inventory.find( name )
+      if ( not object is None ):
+         return object
+      object = self.world.find( name )
+      if ( not object is None ):
+         return object
+      return None
+
+   def use( self, name_of_tool, name_of_actor ):
+      tool = self.inventory.find( name_of_tool )
+      if ( tool is None ):
+         return None
+      actor = self.find( name_of_actor )
+      if ( actor is None ):
+         return None
+      print actor.name
+      print tool.name
+      for action in self.actions:
+         if action.tool == tool.name and action.actor == actor.name:
+            self.destroy( name_of_tool )
+            self.destroy( name_of_actor )
+            self.actions.remove( action )
+            retval = self.world.put( action.prototype )
+            return action.prototype
+      return None
+
+   def is_in_world( self, name ):
+      return self.world.find( name )
 
 class GameObjectAction:
    def __init__( self, actor, tool, actionDescription, prototype ):
